:sparkles: [add] 글 등록 할 때 이미지 첨부 기능 추가
글 등록시 로그인 한 user 의 id 를 주입하여 글 등록을 하는 기능을 추가 하였고
추가로 글 등록 할 때 이미지 파일이 있을 시 이미지 파일을 등록 하는 로직도 추가 하였습니다.

+++ my_project/src/main/java/com/example/demo/dto/Image.java
@@ -10,8 +10,8 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class Image {
-    private Long image_id;
-    private Long user_id;
+    private int image_id;
+    private int user_id;
     private LocalDateTime save_date;
     private String file_name;
     private String file_extention;
